style: moved methods to the right places for better code organization
